GROUNDWORK-2204: Add validations

+++ install.sh
@@ -3,9 +3,8 @@
 TAG=
 DOCKER_USER=
 DOCKER_PASSWORD=
-
-# TODO: use tag
-gw8image="groundworkdevelopment/gw8:GROUNDWORK-2204"
+GW8_INSTANCE_NAME=
+GW8_IMAGE=
 
 __parse_config_yaml() {
   local prefix=$2
@@ -34,32 +33,106 @@ __check_environment() {
     echo "ERROR: 'docker-compose' should be installed"
     exit 1
   fi
-  if ! which psql | grep "psql" >/dev/null 2>/dev/null; then
-    echo "ERROR: 'psql' should be installed"
+
+  DOCKER_VERSION_LIM=20.10.13
+  DOCKER_COMPOSE_VERSION_LIM=1.17.0
+  MSG="ERROR: docker ${DOCKER_VERSION_LIM}+ and docker-compose ${DOCKER_COMPOSE_VERSION_LIM}+ are required"
+  if [ ! "$(command -v docker)" ] || [ ! "$(command -v docker-compose)" ]; then
+    echo "${MSG}"
     exit 1
   fi
-  if ! which influx | grep "influx" >/dev/null 2>/dev/null; then
-    echo "ERROR: 'influx' should be installed"
+  DOCKER_COMPOSE_VERSION=$(docker-compose version --short)
+  DOCKER_VERSION=$(docker version --format '{{.Server.Version}}')
+  if [ "$(__version "$DOCKER_COMPOSE_VERSION")" -lt "$(__version "$DOCKER_COMPOSE_VERSION_LIM")" ] ||
+    [ "$(__version "$DOCKER_VERSION")" -lt "$(__version "$DOCKER_VERSION_LIM")" ]; then
+    echo "${MSG}"
     exit 1
   fi
 }
 
-__set_variables() {
-  if [[ "$version" == "" ]]; then
-    echo "ERROR: 'version' should be specified"
-    exit 1
+__load_alpine() {
+  ## quick tests for failing with separate alpine image before long full load
+  if [ -z "$SKIP_ALPINE" ]; then
+    echo Try to run alpine container
+    if ! docker run --rm alpine:latest echo 'Test alpine OK'; then
+      printf '\nTest alpine FAILED\n'
+      __exit_1
+    fi
+
+    DF=$(docker run --rm alpine:"${ALPINE_VERSION}" sh -c "df -B 1073741824 | awk '/ \/\$/ {printf \"%0.0f\n\",\$4;}'")
+    RAM=$(docker run --rm alpine:"${ALPINE_VERSION}" sh -c "free | awk '/Mem:/ {printf \"%0.0f\n\",\$2/1000000;}'")
+    NPROC=$(docker run --rm alpine:"${ALPINE_VERSION}" nproc)
+    MSG_DF="${DF}GB of disk space available"
+    MSG_RAM="${RAM}GB of RAM installed"
+    MSG_NPROC="${NPROC} cores CPU available"
+    if [ "$NPROC" -eq 1 ]; then
+      MSG_NPROC="Single-core CPU available"
+    fi
+
+    if [ "$DF" -le 24 ]; then
+      MSG=$(
+        cat <<EOM
+ERROR: This system does not meet the minimum requirements
+for GroundWork 8. GroundWork requires a minimum of 25GB
+of disk space to properly install.
+  * 25GB of disk space required - ${MSG_DF}
+EOM
+      )
+      echo "${MSG}"
+      exit 1
+    fi
+
+    if [ "$NPROC" -le 3 ] || [ "$RAM" -le 31 ] || [ "$DF" -le 999 ]; then
+      MSG=$(
+        cat <<EOM
+ERROR: This system does not meet the recommended minimum
+requirements for GroundWork 8. GroundWork recommends the
+following minimum environment specification for correct
+operation in production:
+  * 4 cores CPU - ${MSG_NPROC}
+  * 32GB of RAM - ${MSG_RAM}
+  * 1TB of disk space - ${MSG_DF}
+EOM
+      )
+      echo "${MSG}"
+      exit 1
+    fi
   fi
-  TAG=$version
+}
+
+__set_variables() {
   if [[ "$docker_user" == "" ]]; then
     echo "ERROR: 'docker_user' should be specified"
     exit 1
   fi
   DOCKER_USER=$docker_user
+
   if [[ "$docker_password" == "" ]]; then
     echo "ERROR: 'docker_password' should be specified"
     exit 1
   fi
   DOCKER_PASSWORD=$docker_password
+
+  if [[ "$gw8_tag" == "" ]]; then
+    echo "ERROR: 'gw8_tag' should be specified"
+    exit 1
+  fi
+  TAG=$gw8_tag
+
+  if [[ "$gw8_instance_name" == "" ]]; then
+    echo "ERROR: 'gw8_instance_name' should be specified"
+    exit 1
+  fi
+  GW8_INSTANCE_NAME=$gw8_instance_name
+
+  if [[ "$gw8_image" == "" ]]; then
+    echo "ERROR: 'gw8_image' should be specified"
+    exit 1
+  fi
+  GW8_IMAGE=$gw8_image
+
+  PARENT_INSTANCE_NAME=$gw8_parent_instance_name
+  CHILD_INSTANCE_NAME=$gw8_child_instance_name
 }
 
 __docker_login() {
@@ -67,31 +140,46 @@ __docker_login() {
 }
 
 __pull_gw8_image() {
-  docker pull "$gw8image"
-  # docker pull "$gw8image:${TAG}"
+  docker pull "$GW8_IMAGE:${TAG}"
 }
 
 __extract_gw8_image() {
   mkdir gw8
-  cd gw8 || exit
-  GW8_INSTANCE_NAME=$(hostname -f)
-  PARENT_INSTANCE_NAME=
-  CHILD_INSTANCE_NAME=
-  docker run \
+  if ! cd gw8; then
+    echo "gw8 directory doesn't exist"
+    exit 1
+  fi
+
+  if ! docker run \
     -e GW8_INSTANCE_NAME="$GW8_INSTANCE_NAME" \
     -e PARENT_INSTANCE_NAME="$PARENT_INSTANCE_NAME" \
     -e CHILD_INSTANCE_NAME="$CHILD_INSTANCE_NAME" \
     -v /var/run/docker.sock:/var/run/docker.sock \
     -v "${HOME}"/.docker:/root/.docker \
     -v /tmp:/tmp/tmp \
-    --name gw8 $gw8image /src/docker_cmd.sh
-  # --name gw8 "$gw8image:$TAG" /src/docker_cmd.sh
+    --name gw8 "${GW8_IMAGE}:${TAG}" /src/docker_cmd.sh; then
+    echo "GroundWork 8 Installation FAILED"
+    exit 1
+  fi
+
   docker cp gw8:RELEASE .
-  docker cp gw8:docker-compose.yml .
   docker cp gw8:.env .
-  docker cp gw8:gw8.env .
-  docker cp gw8:docker-compose.override.yml .
+  docker cp gw8:docker-compose.yml .
+
+  if [ ! -f gw8.env ]; then
+    docker cp gw8:gw8.env .
+  fi
+  if [ ! -f docker-compose.override.yml ]; then
+    docker cp gw8:docker-compose.override.yml .
+  fi
   docker rm gw8
+
+  if [ -f "gw8.env" ]; then
+    cmd gw8.env | sed -e '/^\s*TZ\s*=/s/TZ/GW8_TZ/' >gw8.env.updated
+    mv -f gw8.env.updated gw8.env
+    rm -f gw8.env.updated
+  fi
+
   docker-compose pull
   docker-compose up -d
 }
